move log replay provider to new file
